Switched from manual delay to using animation events
Animations now look good because instead of using code to deal damage or throw kunai i am using animation events.
An animation event is putting a function from the code into the actual animation so the kunai is summoned right at the end of the throw animation which makes everything easier and simpler and in general better.